+++ src/main/kotlin/org/mapsAdvisor/mapsAdvisor/service/FeedbackService.kt
@@ -1,10 +1,10 @@
 package org.mapsAdvisor.mapsAdvisor.service
 
-import org.mapsAdvisor.mapsAdvisor.model.Grade
-import org.mapsAdvisor.mapsAdvisor.model.PlaceFeedback
-import org.mapsAdvisor.mapsAdvisor.model.RouteFeedback
-import org.mapsAdvisor.mapsAdvisor.repository.PlaceFeedbackRepository
-import org.mapsAdvisor.mapsAdvisor.repository.RouteFeedbackRepository
+import org.mapsAdvisor.mapsAdvisor.entity.Grade
+import org.mapsAdvisor.mapsAdvisor.entity.PlaceFeedback
+import org.mapsAdvisor.mapsAdvisor.entity.RouteFeedback
+import org.mapsAdvisor.mapsAdvisor.exception.NotFoundException
+import org.mapsAdvisor.mapsAdvisor.repository.*
 import org.mapsAdvisor.mapsAdvisor.request.PlaceFeedbackRequest
 import org.mapsAdvisor.mapsAdvisor.request.RouteFeedbackRequest
 import org.springframework.data.domain.Page
@@ -15,13 +15,18 @@ import org.springframework.stereotype.Service
 @Service
 class FeedbackService(
     private val routeFeedbackRepository: RouteFeedbackRepository,
-    private val placeFeedbackRepository: PlaceFeedbackRepository
+    private val placeFeedbackRepository: PlaceFeedbackRepository,
+    private val routeRepository: RouteRepository,
+    private val placeRepository: PlaceRepository,
+    private val personRepository: PersonRepository,
 ) {
-    companion object {
-        const val SIZE: Int = 50
-    }
-
     fun createRouteFeedback(feedback: RouteFeedbackRequest): RouteFeedback {
+        if (!routeRepository.existsById(feedback.routeId)) {
+            throw NotFoundException("Route with id ${feedback.routeId} not found")
+        }
+        if (!personRepository.existsById(feedback.personId)) {
+            throw NotFoundException("Person with id ${feedback.personId} not found")
+        }
         return routeFeedbackRepository.save(
             RouteFeedback(
                 routeId = feedback.routeId,
@@ -33,19 +38,31 @@ class FeedbackService(
         )
     }
 
-    fun getRouteFeedbacks(routeId: String, page: Int): Page<RouteFeedback> {
-        val pageable: Pageable = PageRequest.of(page, SIZE)
+    fun getRouteFeedbacks(routeId: String, page: Int, size: Int): Page<RouteFeedback> {
+        if (!routeRepository.existsById(routeId)) {
+            throw NotFoundException("Route with id $routeId not found")
+        }
+        val pageable: Pageable = PageRequest.of(page, size)
         return routeFeedbackRepository.findByRouteId(routeId, pageable)
     }
 
     fun deleteRouteFeedback(feedbackId: String) {
+        if (!routeFeedbackRepository.existsById(feedbackId)) {
+            throw NotFoundException("Feedback with id $feedbackId not found")
+        }
         routeFeedbackRepository.deleteById(feedbackId)
     }
 
     fun createPlaceFeedback(feedback: PlaceFeedbackRequest): PlaceFeedback {
+        if (!placeRepository.existsById(feedback.placeId)) {
+            throw NotFoundException("Place with id ${feedback.placeId} not found")
+        }
+        if (!personRepository.existsById(feedback.personId)) {
+            throw NotFoundException("Person with id ${feedback.personId} not found")
+        }
         return placeFeedbackRepository.save(
             PlaceFeedback(
-                placeId = feedback.routeId,
+                placeId = feedback.placeId,
                 grade = Grade(
                     personId = feedback.personId,
                     grade = feedback.grade,
@@ -54,12 +71,18 @@ class FeedbackService(
         )
     }
 
-    fun getPlaceFeedbacks(routeId: String, page: Int): Page<PlaceFeedback> {
-        val pageable: Pageable = PageRequest.of(page, SIZE)
-        return placeFeedbackRepository.findByPlaceId(routeId, pageable)
+    fun getPlaceFeedbacks(placeId: String, page: Int, size: Int): Page<PlaceFeedback> {
+        if (!placeFeedbackRepository.existsById(placeId)) {
+            throw NotFoundException("Place with id $placeId not found")
+        }
+        val pageable: Pageable = PageRequest.of(page, size)
+        return placeFeedbackRepository.findByPlaceId(placeId, pageable)
     }
 
     fun deletePlaceFeedback(feedbackId: String) {
+        if (!placeFeedbackRepository.existsById(feedbackId)) {
+            throw NotFoundException("Feedback with id $feedbackId not found")
+        }
         placeFeedbackRepository.deleteById(feedbackId)
     }
